First feeble attempt at gtk module. Also /* */ comments and misc additions

+++ type.h
@@ -7,8 +7,8 @@
 
 // TODO review size_t and so on, ssize_t is signed size
 
-typedef unsigned long int index_t; // TODO 64 bit
-typedef long int integer_t; // TODO 64 bit
+typedef unsigned long long int index_t; // TODO 64 bit
+typedef long long int integer_t; // TODO 64 bit
 typedef char char_t; // TODO unsigned
 
 #define TYPE_H
